cleanup: remove structs, re arrange codE
