+++ crates/rust-client/src/store/sqlite_store/transaction.rs
@@ -16,15 +16,11 @@ use miden_objects::{
         ExecutedTransaction, OutputNotes, ToInputNoteCommitments, TransactionId, TransactionScript,
     },
 };
-use rusqlite::{params, types::Value};
+use rusqlite::{Connection, Transaction, params, types::Value};
 use tracing::info;
 
 use super::{
-    SqliteStore,
-    account::update_account,
-    db_management::{connection::Connection, transaction::Transaction},
-    note::apply_note_updates_tx,
-    sync::add_note_tag_tx,
+    SqliteStore, account::update_account, note::apply_note_updates_tx, sync::add_note_tag_tx,
 };
 use crate::{
     insert_sql,
